pyproject.toml
# Copyright (c) Meta Platforms, Inc. and affiliates.
# All rights reserved.
#
# This source code is licensed under the BSD-style license found in the
# LICENSE file in the root directory of this source tree.

[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "openenv-cloudscale_RL"
version = "0.1.0"
description = "Cloudscale Rl environment for OpenEnv with LLM multi-agent orchestration"
requires-python = ">=3.10"
dependencies = [
    # Core OpenEnv runtime (provides FastAPI server + HTTP client types)
    "openenv-core[core]>=0.2.2",
    
    # Environment-specific dependencies we added for your project
    "pydantic>=2.0.0",      # Required for OpenEnv models
    "gradio>=4.0.0",        # For app_dashboard.py
    "matplotlib>=3.7.0",    # For telemetry charts in dashboard
    "anthropic>=0.18.0",    # For the LLM AutoScalerAgent
    "requests>=2.31.0",     # For Ollama API calls in AutoScalerAgent
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-cov>=4.0.0",
]

[project.scripts]
# Server entry point - enables running via: uv run --project . server
# or: python -m cloudscale_RL.server.app
server = "cloudscale_RL.server.app:main"

[tool.setuptools]
include-package-data = true
packages = ["cloudscale_RL", "cloudscale_RL.server"]
package-dir = { "cloudscale_RL" = ".", "cloudscale_RL.server" = "server" }